Changing the Safeguard Page to css modules
