--- src/components/SafeguardBottom.jsx
@@ -1,23 +1,0 @@
-function SafeguardBottom() {
-    return(
-        <>
-         {/* Contact Support Section */}
-        <div className="mt-8 sm:mt-12 text-center">
-          <h2 className="text-xl sm:text-2xl font-semibold text-gray-900 mb-4 sm:mb-6">
-            Need to talk or unsure where to start?
-          </h2>
-          
-          <p className="text-sm sm:text-base text-gray-600 leading-relaxed mb-6 sm:mb-8 max-w-2xl mx-auto">
-            The resources above are here to support you. If you need further guidance or prefer to 
-            speak with someone from CompLet101™ directly, please reach out to our support team.
-          </p>
-          
-          <button className="bg-[#2677B2] hover:bg-[#1f5f8f] text-white font-medium px-6 py-3 rounded-lg text-sm sm:text-base transition-colors duration-200">
-            Contact CompLet101™ Support
-          </button>
-
-        </div>
-        </>
-    )
-}
-export default SafeguardBottom;
